+++ python/tests/test_array_api.py
@@ -33,14 +33,14 @@ def test_to_source(snapshot_py):
     _NDArray_1 = NDArray.var("X")
     X_orig = copy(_NDArray_1)
     assume_dtype(_NDArray_1, DType.float64)
-    assume_shape(_NDArray_1, TupleInt(Int(150)) + TupleInt(Int(4)))
+    # assume_shape(_NDArray_1, TupleInt(Int(150)) + TupleInt(Int(4)))
 
     _NDArray_2 = NDArray.var("y")
     Y_orig = copy(_NDArray_2)
 
     assume_dtype(_NDArray_2, int64)
     assume_shape(_NDArray_2, TupleInt(150))  # type: ignore
-    assume_value_one_of(_NDArray_2, (0, 1, 2))  # type: ignore
+    # assume_value_one_of(_NDArray_2, (0, 1, 2))  # type: ignore
 
     _NDArray_3 = reshape(_NDArray_2, TupleInt(Int(-1)))
     _NDArray_4 = astype(unique_counts(_NDArray_3)[Int(1)], DType.float64) / NDArray.scalar(Value.float(Float(150.0)))
@@ -51,27 +51,24 @@ def test_to_source(snapshot_py):
     )
     _MultiAxisIndexKey_1 = MultiAxisIndexKey(MultiAxisIndexKeyItem.slice(Slice()))
     _IndexKey_1 = IndexKey.multi_axis(MultiAxisIndexKey(MultiAxisIndexKeyItem.int(Int(0))) + _MultiAxisIndexKey_1)
-    _NDArray_5[_IndexKey_1] = mean(
-        _NDArray_1[ndarray_index(unique_inverse(_NDArray_3)[Int(1)] == NDArray.scalar(Value.int(Int(0))))],
-        OptionalIntOrTuple.int(Int(0)),
-    )
-    _IndexKey_2 = IndexKey.multi_axis(MultiAxisIndexKey(MultiAxisIndexKeyItem.int(Int(1))) + _MultiAxisIndexKey_1)
-    _NDArray_5[_IndexKey_2] = mean(
-        _NDArray_1[ndarray_index(unique_inverse(_NDArray_3)[Int(1)] == NDArray.scalar(Value.int(Int(1))))],
-        OptionalIntOrTuple.int(Int(0)),
-    )
-    _IndexKey_3 = IndexKey.multi_axis(MultiAxisIndexKey(MultiAxisIndexKeyItem.int(Int(2))) + _MultiAxisIndexKey_1)
-    _NDArray_5[_IndexKey_3] = mean(
-        _NDArray_1[ndarray_index(unique_inverse(_NDArray_3)[Int(1)] == NDArray.scalar(Value.int(Int(2))))],
-        OptionalIntOrTuple.int(Int(0)),
-    )
+    _NDArray_5 = _NDArray_1[ndarray_index(unique_inverse(_NDArray_3)[Int(1)] == NDArray.scalar(Value.int(Int(0))))]
+    # _IndexKey_2 = IndexKey.multi_axis(MultiAxisIndexKey(MultiAxisIndexKeyItem.int(Int(1))) + _MultiAxisIndexKey_1)
+    # _NDArray_5[_IndexKey_2] = mean(
+    #     _NDArray_1[ndarray_index(unique_inverse(_NDArray_3)[Int(1)] == NDArray.scalar(Value.int(Int(1))))],
+    #     OptionalIntOrTuple.int(Int(0)),
+    # )
+    # _IndexKey_3 = IndexKey.multi_axis(MultiAxisIndexKey(MultiAxisIndexKeyItem.int(Int(2))) + _MultiAxisIndexKey_1)
+    # _NDArray_5[_IndexKey_3] = mean(
+    #     _NDArray_1[ndarray_index(unique_inverse(_NDArray_3)[Int(1)] == NDArray.scalar(Value.int(Int(2))))],
+    #     OptionalIntOrTuple.int(Int(0)),
+    # )
 
-    res = _NDArray_3 + _NDArray_5 + _NDArray_4
+    res = _NDArray_1[ndarray_index(_NDArray_2)]
     egraph = EGraph([array_api_module])
     egraph.register(res)
     egraph.run(10000)
     res = egraph.extract(res)
-
+    # print(array_api_module.unextractable().as_egglog_string)
     egraph = EGraph([array_api_module_string])
     fn = ndarray_program(res).function_two(ndarray_program(X_orig), ndarray_program(Y_orig))
     egraph.register(fn)
@@ -80,11 +77,12 @@ def test_to_source(snapshot_py):
 
     egraph = EGraph([array_api_module_string])
     egraph.register(fn.eval_py_object(egraph.save_object({"np": numpy})))
-    egraph.run(1000000)
+    while egraph.run(1).updated:
+        pass
     # print(egraph.run(10))
     # egraph.display(n_inline_leaves=0, split_primitive_outputs=True)
-    fn_source = egraph.load_object(egraph.extract(PyObject.from_string(fn.statements)))
-    assert fn_source == snapshot_py
+    # fn_source = egraph.load_object(egraph.extract(PyObject.from_string(fn.statements)))
+    # assert fn_source == snapshot_py
 
 
 # @pytest.mark.xfail(raises=TODO)