+++ docs/changelog.md
@@ -4,7 +4,10 @@ _This project uses semantic versioning. Before 1.0.0, this means that every brea
 
 ## Unreleased
 
-...
+### New Features
+
+- Add ability to print egglog string of module with `.as_egglog_string`
+- Adds ability to visualize changes in egraph as it runs with `.saturate()`
 
 ## 0.7.0 (2023-10-04)